backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
anthropic==0.42.0
openai==1.58.1
google-generativeai==0.8.3
httpx==0.28.1
python-multipart==0.0.20
jinja2==3.1.5
aiofiles==24.1.0
websockets==14.1
cryptography==44.0.0
pydantic==2.10.4
pydantic-settings==2.7.1
weasyprint==63.1
tiktoken==0.8.0
